2841: make KDL lib static, allows to hide STL includes from string utilities
