+++ common/src/Assets/EntityDefinitionFileSpec.cpp
@@ -21,7 +21,7 @@
 
 #include "StringUtils.h"
 
-#include <kdl/string_utils.h>
+#include <kdl/string_compare.h>
 
 #include <cassert>
