+++ common/src/IO/FgdParser.cpp
@@ -26,7 +26,7 @@
 #include "IO/ELParser.h"
 #include "IO/LegacyModelDefinitionParser.h"
 
-#include <kdl/string_utils.h>
+#include <kdl/string_compare.h>
 #include <kdl/vector_utils.h>
 
 #include <memory>
