+++ common/src/Model/TagMatcher.cpp
@@ -31,7 +31,7 @@
 #include "Model/NodeCollection.h"
 #include "Model/World.h"
 
-#include <kdl/string_utils.h>
+#include <kdl/string_compare.h>
 #include <kdl/vector_utils.h>
 
 #include <vector>
