+++ common/src/View/MapViewBase.cpp
@@ -61,7 +61,7 @@
 #include "View/SelectionTool.h"
 #include "View/QtUtils.h"
 
-#include <kdl/string_utils.h>
+#include <kdl/string_compare.h>
 
 #include <vecmath/util.h>
