+++ common/src/View/SmartAttributeEditorMatcher.cpp
@@ -19,7 +19,7 @@
 
 #include "SmartAttributeEditorMatcher.h"
 
-#include <kdl/string_utils.h>
+#include <kdl/string_compare.h>
 
 #include <set>
 #include <vector>
