+++ common/test/src/IO/DefParserTest.cpp
@@ -31,7 +31,7 @@
 #include "IO/TestParserStatus.h"
 #include "TestUtils.h"
 
-#include <kdl/string_utils.h>
+#include <kdl/string_compare.h>
 #include <kdl/vector_utils.h>
 
 namespace TrenchBroom {
