+++ common/test/src/IO/NodeWriterTest.cpp
@@ -30,7 +30,7 @@
 #include "Model/MapFormat.h"
 #include "Model/World.h"
 
-#include <kdl/string_utils.h>
+#include <kdl/string_compare.h>
 
 #include <vector>
