+++ lib/kdl/CMakeLists.txt
@@ -8,24 +8,27 @@ if (POLICY CMP0092)
 endif()
 
 set(KDL_INCLUDE_DIR ${CMAKE_CURRENT_SOURCE_DIR}/include)
+set(KDL_SOURCE_DIR ${CMAKE_CURRENT_SOURCE_DIR}/src)
 
-add_library(kdl INTERFACE)
-target_compile_features(kdl INTERFACE cxx_std_17)
-target_include_directories(kdl INTERFACE
-        $<BUILD_INTERFACE:${KDL_INCLUDE_DIR}>
-        $<INSTALL_INTERFACE:kdl/include/kdl>)
+set(KDL_SOURCE
+        "${KDL_SOURCE_DIR}/string_compare.cpp")
+
+add_library(kdl STATIC ${KDL_SOURCE})
+target_compile_features(kdl PUBLIC cxx_std_17)
+target_include_directories(kdl
+        PUBLIC ${KDL_INCLUDE_DIR})
 
 if(CMAKE_CXX_COMPILER_ID STREQUAL "Clang" OR CMAKE_CXX_COMPILER_ID STREQUAL "AppleClang")
-    target_compile_options(kdl INTERFACE -Wall -Wextra -pedantic -Wno-c++98-compat -Wno-padded)
+    target_compile_options(kdl PUBLIC -Wall -Wextra -pedantic -Wno-c++98-compat -Wno-padded)
 elseif(CMAKE_CXX_COMPILER_ID STREQUAL "GNU")
-    target_compile_options(kdl INTERFACE -Wall -Wextra -pedantic)
+    target_compile_options(kdl PUBLIC -Wall -Wextra -pedantic)
 elseif(MSVC EQUAL 1)
-    target_compile_options(kdl INTERFACE /W4 /EHsc /MP)
+    target_compile_options(kdl PUBLIC /W4 /EHsc /MP)
 
     # signed/unsigned mismatch: https://docs.microsoft.com/en-us/cpp/error-messages/compiler-warnings/compiler-warning-level-4-c4365
-    target_compile_options(kdl INTERFACE /w44365)
+    target_compile_options(kdl PUBLIC /w44365)
 else()
-    message(FATAL_ERROR "Cannot set compile options for target common")
+    message(FATAL_ERROR "Cannot set compile options for target")
 endif()
 
 # Prevent overriding the parent project's compiler/linker
