+++ lib/kdl/include/kdl/collection_utils.h
@@ -238,6 +238,19 @@ namespace kdl {
             return kdl::range_is_equivalent(std::begin(c1), std::end(c1), std::begin(c2), std::end(c2), cmp);
         }
     }
+
+    /**
+     * Sorts the elements of the given collection according to the given comparator.
+     *
+     * @tparam C the collection type
+     * @tparam Compare the type of the comparator to use
+     * @param c the collection to sort
+     * @param cmp the comparator to use for comparisons
+     */
+    template<typename C, typename Compare = std::less<typename C::value_type>>
+    void sort(C& c, const Compare& cmp = Compare()) {
+        std::sort(std::begin(c), std::end(c), cmp);
+    }
 }
 
 #endif //KDL_COLLECTION_UTILS_H
