+++ lib/kdl/test/CMakeLists.txt
@@ -4,7 +4,7 @@ target_sources(kdl-test PRIVATE
         "${CMAKE_CURRENT_SOURCE_DIR}/src/map_utils_test.cpp"
         "${CMAKE_CURRENT_SOURCE_DIR}/src/run_all.cpp"
         "${CMAKE_CURRENT_SOURCE_DIR}/src/set_adapter_test.cpp"
-        "${CMAKE_CURRENT_SOURCE_DIR}/src/string_utils_test.cpp"
+        "${CMAKE_CURRENT_SOURCE_DIR}/src/string_compare_test.cpp"
         "${CMAKE_CURRENT_SOURCE_DIR}/src/vector_set_test.cpp"
         "${CMAKE_CURRENT_SOURCE_DIR}/src/vector_utils_test.cpp"
         )
